refactor(api): separate route sync responsibilities
Extract route persistence and sync checkpoint handling from RoutesSyncService, leaving it responsible for orchestration. Add focused unit tests for each service boundary.
