+++ apps/api/src/sync/route-persistence.service.spec.ts
@@ -1,0 +1,130 @@
+import { CityNameType, DirectionType } from '@bus/shared'
+import type { TdxBusRoute } from '@bus/shared'
+import type { PrismaService } from '../prisma/prisma.service.js'
+import { routeMapper } from './mappers/route.mapper.js'
+import { RoutePersistenceService } from './route-persistence.service.js'
+
+const tdxRoute: TdxBusRoute = {
+  RouteUID: 'NWT10116',
+  RouteID: '10116',
+  HasSubRoutes: true,
+  Operators: [
+    {
+      OperatorID: '16176',
+      OperatorName: {
+        Zh_tw: '臺北客運',
+        En: 'Taipei Bus Co., Ltd.',
+      },
+      OperatorNo: '0400',
+    },
+  ],
+  AuthorityID: '005',
+  ProviderID: '015',
+  SubRoutes: [
+    {
+      SubRouteUID: 'NWT101160',
+      SubRouteID: '101160',
+      OperatorIDs: ['16176'],
+      SubRouteName: { Zh_tw: '242', En: '242' },
+      Direction: DirectionType.GO,
+      FirstBusTime: '0530',
+      LastBusTime: '1710',
+    },
+  ],
+  BusRouteType: 11,
+  RouteName: { Zh_tw: '242', En: '242' },
+  DepartureStopNameZh: '中和',
+  DepartureStopNameEn: 'Zhonghe',
+  DestinationStopNameZh: '西門',
+  DestinationStopNameEn: 'Ximen',
+  UpdateTime: '2026-06-03T18:45:21+08:00',
+  VersionID: 7931,
+}
+
+function createPersistenceMocks() {
+  const calls = {
+    routeUpsert: [] as unknown[],
+    subRouteUpsert: [] as unknown[],
+    operatorUpsert: [] as unknown[],
+    routeOperatorUpsert: [] as unknown[],
+  }
+  const prismaService = {
+    route: {
+      findMany: () => Promise.resolve([]),
+      upsert: (args: unknown) => {
+        calls.routeUpsert.push(args)
+        return Promise.resolve({ id: 'route-db-id' })
+      },
+      updateMany: () => Promise.resolve({ count: 2 }),
+    },
+    subRoute: {
+      upsert: (args: unknown) => {
+        calls.subRouteUpsert.push(args)
+        return Promise.resolve({ id: 'subroute-db-id' })
+      },
+      updateMany: () => Promise.resolve({ count: 0 }),
+    },
+    operator: {
+      upsert: (args: unknown) => {
+        calls.operatorUpsert.push(args)
+        return Promise.resolve({ id: 'operator-db-id' })
+      },
+    },
+    routeOperator: {
+      upsert: (args: unknown) => {
+        calls.routeOperatorUpsert.push(args)
+        return Promise.resolve({})
+      },
+      deleteMany: () => Promise.resolve({ count: 0 }),
+    },
+  }
+
+  return { calls, prismaService }
+}
+
+describe('RoutePersistenceService', () => {
+  it('persists routes with their subroutes and operators', async () => {
+    const { calls, prismaService } = createPersistenceMocks()
+    const service = new RoutePersistenceService(
+      prismaService as unknown as PrismaService,
+    )
+    const routes = routeMapper({
+      city: CityNameType.NEW_TAIPEI,
+      tdxRoutes: [tdxRoute],
+    })
+
+    await expect(
+      service.persistRoutes(routes, { city: CityNameType.NEW_TAIPEI }),
+    ).resolves.toEqual({
+      records_read: 1,
+      records_created: 1,
+      records_updated: 0,
+      records_deactivated: 2,
+    })
+
+    expect(calls.routeUpsert).toEqual([
+      expect.objectContaining({ where: { uuid: 'NWT10116' } }),
+    ])
+    expect(calls.subRouteUpsert).toEqual([
+      expect.objectContaining({ where: { uuid: 'NWT101160-0' } }),
+    ])
+    expect(calls.operatorUpsert).toEqual([
+      expect.objectContaining({ where: { tdx_operator_id: '16176' } }),
+    ])
+    expect(calls.routeOperatorUpsert).toEqual([
+      {
+        where: {
+          route_id_operator_id: {
+            route_id: 'route-db-id',
+            operator_id: 'operator-db-id',
+          },
+        },
+        create: {
+          route_id: 'route-db-id',
+          operator_id: 'operator-db-id',
+        },
+        update: {},
+      },
+    ])
+  })
+})
