+++ apps/api/src/sync/routes-sync.service.spec.ts
@@ -1,185 +1,60 @@
-import { CityNameType, DirectionType, getEnumValues } from '@bus/shared'
-import type { TdxBusRoute } from '@bus/shared'
-import { SyncStatusType as PrismaSyncStatusType } from '../generated/prisma/enums.js'
-import type { PrismaService } from '../prisma/prisma.service.js'
+import { CityNameType, getEnumValues } from '@bus/shared'
 import { cityMapper } from './mappers/route.mapper.js'
+import type { RoutePersistenceService } from './route-persistence.service.js'
 import { RoutesSyncService } from './routes-sync.service.js'
+import type { SyncCheckpointService } from './sync-checkpoint.service.js'
 import type { TdxClientService } from './tdx-client.service.js'
 import { TdxMonthlyQuotaExceededError } from './tdx-client.service.js'
 
-const tdxRoute: TdxBusRoute = {
-  RouteUID: 'NWT10116',
-  RouteID: '10116',
-  HasSubRoutes: true,
-  Operators: [
-    {
-      OperatorID: '16176',
-      OperatorName: {
-        Zh_tw: '臺北客運',
-        En: 'Taipei Bus Co., Ltd.',
-      },
-      OperatorNo: '0400',
-    },
-  ],
-  AuthorityID: '005',
-  ProviderID: '015',
-  SubRoutes: [
-    {
-      SubRouteUID: 'NWT101160',
-      SubRouteID: '101160',
-      OperatorIDs: ['16176'],
-      SubRouteName: { Zh_tw: '242', En: '242' },
-      Direction: DirectionType.GO,
-      FirstBusTime: '0530',
-      LastBusTime: '1710',
-    },
-  ],
-  BusRouteType: 11,
-  RouteName: { Zh_tw: '242', En: '242' },
-  DepartureStopNameZh: '中和',
-  DepartureStopNameEn: 'Zhonghe',
-  DestinationStopNameZh: '西門',
-  DestinationStopNameEn: 'Ximen',
-  UpdateTime: '2026-06-03T18:45:21+08:00',
-  VersionID: 7931,
-}
-
-function createPersistenceMocks() {
-  const calls = {
-    routeUpsert: [] as unknown[],
-    subRouteUpsert: [] as unknown[],
-    operatorUpsert: [] as unknown[],
-    routeOperatorUpsert: [] as unknown[],
-  }
-  const prismaService = {
-    route: {
-      findMany: () => Promise.resolve([]),
-      upsert: (args: unknown) => {
-        calls.routeUpsert.push(args)
-        return Promise.resolve({ id: 'route-db-id' })
-      },
-      updateMany: () => Promise.resolve({ count: 2 }),
-    },
-    subRoute: {
-      upsert: (args: unknown) => {
-        calls.subRouteUpsert.push(args)
-        return Promise.resolve({ id: 'subroute-db-id' })
-      },
-      updateMany: () => Promise.resolve({ count: 0 }),
-    },
-    operator: {
-      upsert: (args: unknown) => {
-        calls.operatorUpsert.push(args)
-        return Promise.resolve({ id: 'operator-db-id' })
-      },
-    },
-    routeOperator: {
-      upsert: (args: unknown) => {
-        calls.routeOperatorUpsert.push(args)
-        return Promise.resolve({})
-      },
-      deleteMany: () => Promise.resolve({ count: 0 }),
-    },
-    syncRun: {
-      update: () => Promise.resolve({}),
-    },
-    syncRunCity: {
-      update: () => Promise.resolve({}),
-    },
-  }
-
-  return { calls, prismaService }
-}
-
 describe('RoutesSyncService', () => {
-  it('maps and persists routes with their subroutes and operators', async () => {
-    const { calls, prismaService } = createPersistenceMocks()
-    const tdxClientService = {
-      fetchRoutes: () => Promise.resolve([tdxRoute]),
-    }
-    const service = new RoutesSyncService(
-      prismaService as unknown as PrismaService,
-      tdxClientService as unknown as TdxClientService,
-    )
-
-    await expect(
-      service.syncCityRoutes(CityNameType.NEW_TAIPEI, 'sync-run-id'),
-    ).resolves.toEqual({
-      records_read: 1,
-      records_created: 1,
-      records_updated: 0,
-      records_deactivated: 2,
-    })
-
-    expect(calls.routeUpsert).toEqual([
-      expect.objectContaining({ where: { uuid: 'NWT10116' } }),
-    ])
-    expect(calls.subRouteUpsert).toEqual([
-      expect.objectContaining({ where: { uuid: 'NWT101160-0' } }),
-    ])
-    expect(calls.operatorUpsert).toEqual([
-      expect.objectContaining({ where: { tdx_operator_id: '16176' } }),
-    ])
-    expect(calls.routeOperatorUpsert).toEqual([
-      {
-        where: {
-          route_id_operator_id: {
-            route_id: 'route-db-id',
-            operator_id: 'operator-db-id',
-          },
-        },
-        create: {
-          route_id: 'route-db-id',
-          operator_id: 'operator-db-id',
-        },
-        update: {},
-      },
-    ])
-  })
-
-  it('marks a sync run pending when the monthly quota is exhausted', async () => {
+  it('marks the current city and run when the monthly quota is exhausted', async () => {
     const retryAt = new Date('2026-07-01T00:00:00.000Z')
     const quotaError = new TdxMonthlyQuotaExceededError(
       'TDX monthly request quota has been exhausted.',
       retryAt,
     )
-    const syncRunUpdates: unknown[] = []
-    const prismaService = {
-      syncRun: {
-        update: (args: unknown) => {
-          syncRunUpdates.push(args)
-          return Promise.resolve({})
-        },
-      },
-      syncRunCity: {
-        createMany: () => Promise.resolve({ count: 22 }),
-        findMany: () => Promise.resolve([]),
-        update: () => Promise.resolve({}),
-      },
-    }
+    const failedCities: unknown[] = []
+    const failedRuns: unknown[] = []
     const tdxClientService = {
       fetchRoutes: () => Promise.reject(quotaError),
     }
+    const routePersistenceService = {}
+    const syncCheckpointService = {
+      startRun: () => Promise.resolve(),
+      ensureCities: () => Promise.resolve(),
+      getCompletedCities: () => Promise.resolve(new Map()),
+      startCity: () => Promise.resolve(),
+      failCity: (...args: unknown[]) => {
+        failedCities.push(args)
+        return Promise.resolve()
+      },
+      failRun: (...args: unknown[]) => {
+        failedRuns.push(args)
+        return Promise.resolve()
+      },
+    }
     const service = new RoutesSyncService(
-      prismaService as unknown as PrismaService,
       tdxClientService as unknown as TdxClientService,
+      routePersistenceService as RoutePersistenceService,
+      syncCheckpointService as unknown as SyncCheckpointService,
     )
 
     await expect(service.syncAllRoutes('sync-run-id')).rejects.toBe(quotaError)
 
-    expect(syncRunUpdates).toEqual([
-      expect.anything(),
-      {
-        where: { id: 'sync-run-id' },
-        data: {
-          status: PrismaSyncStatusType.PENDING,
-          resume_after_at: retryAt,
+    expect(failedCities).toEqual([
+      ['sync-run-id', CityNameType.TAIPEI, quotaError],
+    ])
+    expect(failedRuns).toEqual([
+      [
+        'sync-run-id',
+        quotaError,
+        {
           records_read: 0,
           records_created: 0,
           records_updated: 0,
           records_deactivated: 0,
         },
-      },
+      ],
     ])
   })
 
@@ -192,24 +67,28 @@ describe('RoutesSyncService', () => {
       records_updated: 0,
       records_deactivated: 0,
     }))
-    const prismaService = {
-      syncRun: {
-        update: () => Promise.resolve({}),
-      },
-      syncRunCity: {
-        createMany: () => Promise.resolve({ count: 0 }),
-        findMany: () => Promise.resolve(checkpoints),
-      },
-    }
     const tdxClientService = {
       fetchRoutes: () => {
         fetchCount += 1
         return Promise.resolve([])
       },
     }
+    const routePersistenceService = {}
+    const syncCheckpointService = {
+      startRun: () => Promise.resolve(),
+      ensureCities: () => Promise.resolve(),
+      getCompletedCities: () =>
+        Promise.resolve(
+          new Map(
+            checkpoints.map((checkpoint) => [checkpoint.city, checkpoint]),
+          ),
+        ),
+      completeRun: () => Promise.resolve(),
+    }
     const service = new RoutesSyncService(
-      prismaService as unknown as PrismaService,
       tdxClientService as unknown as TdxClientService,
+      routePersistenceService as RoutePersistenceService,
+      syncCheckpointService as unknown as SyncCheckpointService,
     )
 
     await expect(service.syncAllRoutes('sync-run-id')).resolves.toEqual({
